First sketch of the fundamental frequency detection algorithm
The picking or plucking of a string is detected by keeping track of the volume.
A volume change above a certain threshold triggers the broad scan. Some samples
after the trigger need to be discarded due to harmonics present immediately after
picking/plucking a string. After those samples a batch of samples is taken and
their fundamental period length is determined. The median of that batch of samples
is then calculated to focus on a smaller range of period lengths where no filtering
is required. Once the volume goes below another threshold the focussed scan stops to
avoid displaying spurious results.